Allow using localhost if useNetworkGateway=false
This commit adds a boolean flag "useNetworkGateway" that, when
set to true (default), will use the current method of accessing
the service ports through the gateway, but when set to false
will use localhost instead.

This allows the usage of the plugin on the new Docker for Mac
beta, where exposed ports are accessible on localhost only.